% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageIntro.R
\docType{package}
\name{InitPack}
\alias{InitPack}
\alias{InitPack-package}
\title{InitPack: The first attempt to create a package}
\description{
The InitPack is just the first attempt to create a package in R
}
\section{See Also}{

Useful links:
}

